+++ R/raster2Graph.R
@@ -1,35 +1,38 @@
 
 #' Create a graph from an raster according the connection between cells
 #'
-#' @param r  an object of raster. The value of each cell of the raster is the ‘smoothness’ to indicate how easy the cell connecting with neighbor cells.
-#' @param res Numeric. Resample the input raster to given resolution and use the resampled raster to build graph. Set this to NULL if using the original resolution of of the input raster.
+#' @param r  An object of stars or RasterLayer. The value of each cell of the raster is the ‘smoothness’ to indicate how easy the cell connecting with neighbor cells.
+#' @param cellsize Numeric. Re-sample the input raster to given resolution and use the re-sampled raster to build graph. Set this to NULL if using the original resolution of of the input raster.
 #' @param silent Boolean. A logical indicating if some “progress report” should be given. Default is TRUE.
 #'
-#' @return a list with an graph and the resampled raster. The graph is igraph object, with cells as node and connections as weight
+#' @return a list with an graph and the re-sampled raster (a object of stars). The graph is igraph object, with cells as node and connections as weight.
 #' @export
 #'
 #' @examples
 #' # read in habitat suitability data of wolf in Europe
-#' library(raster)
+#' library(stars)
 #' hsi.file = system.file("extdata","wolf3_int.tif",package="habCluster")
-#' wolf = raster(hsi.file)
+#' wolf = read_stars(hsi.file)
 #' # build graph from raster
 #' g = raster2Graph(wolf, 40000)
 
-raster2Graph  <- function(r, res=NULL,silent=TRUE){
+raster2Graph  <- function(r, cellsize=NULL,silent=TRUE){
+
+  if(is(r,"RasterLayer")){
+    r = stars::st_as_stars(r)
+  }
 
   r2 = NULL
-  if(!is.null(res)){
-    mask = raster::raster(ext = raster::extent(r), crs = raster::crs(r), res = res)
+  if(!is.null(cellsize)){
     if(!silent){
     cat('\nresampling...')
     }
-    r2 = raster::resample(r, mask)
+    r2 = stars::st_warp(r, crs=sf::st_crs(r), cellsize = cellsize)
   }else{
     r2 = r
   }
 
-  matrix = raster::as.matrix(r2)
+  matrix = r2[[1]]
   if(!silent){
   cat('\nextracting edges...')
   }