+++ tests/testthat/test-cluster.R
@@ -1,14 +1,16 @@
 
 test_that("cluster method", {
-  require(raster)
+  library(habCluster)
+  require(stars)
+  require(sf)
   # read data
   hsi.file = system.file("extdata","wolf3_int.tif",package="habCluster")
-  wolf = raster(hsi.file)
+  wolf = read_stars(hsi.file)
 
-  clst = cluster(wolf, method = cluster_leiden, res = 80000, rp = 0.02)
+  clst = cluster(wolf, method = cluster_leiden, cellsize = 80000, rp = 0.02)
 
   # check results are right objects
-  testthat::expect_s4_class(clst$boundary,"SpatialPolygonsDataFrame")
+  testthat::expect_s3_class(clst$boundary,"sf")
   testthat::expect_s3_class(clst$communities ,"communities")
 
 })