+++ tests/testthat/test-raster2Graph.R
@@ -1,13 +1,13 @@
 
 test_that("raster2Graph method", {
-  require(raster)
+  require(stars)
   # read data
   hsi.file = system.file("extdata","wolf3_int.tif",package="habCluster")
-  wolf = raster(hsi.file)
+  wolf = read_stars(hsi.file)
 
   g = raster2Graph(wolf, 80000)
 
   # check results are right objects
-  testthat::expect_s4_class(g$raster,"RasterLayer")
+  testthat::expect_s3_class(g$raster,"stars")
   testthat::expect_s3_class(g$graph ,"igraph")
 })